# Conflicts:
#	src/main/java/pl/kuba/api/apies/BranchManagementAPI.java
#	src/main/java/pl/kuba/api/apies/RentalCompanyAPI.java
#	src/main/java/pl/kuba/api/request/branch/BranchManagementFireRequest.java
#	src/main/java/pl/kuba/api/request/branch/BranchManagementHireRequest.java
#	src/main/java/pl/kuba/domain/servises/BranchManagementService.java
#	src/main/java/pl/kuba/domain/servises/CarService.java
#	src/main/java/pl/kuba/domain/servises/RentalCompanyService.java
#	src/main/java/pl/kuba/domain/servises/ReservationService.java
#	src/main/java/pl/kuba/domain/servises/RevenueService.java
#	src/main/java/pl/kuba/infrastructure/datehelpers/DateFormatter.java
#	src/main/java/pl/kuba/infrastructure/datehelpers/StringToDateConverter.java
#	src/main/java/pl/kuba/infrastructure/persistence/BranchRepository.java
#	src/main/java/pl/kuba/infrastructure/persistence/WorkerRepository.java

+++ src/main/java/pl/kuba/infrastructure/persistence/BranchRepository.java
@@ -5,6 +5,5 @@ import org.springframework.stereotype.Repository;
 import pl.kuba.entities.Branch;
 
 @Repository
-public interface BranchRepository extends JpaRepository<Branch, Integer> {
-
+public interface BranchRepository extends JpaRepository<Branch, Long> {
 }
